+++ starsky/starsky.feature.rename/Services/RenameService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using starsky.foundation.database.Helpers;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Models;
@@ -8,6 +9,7 @@ using starsky.foundation.platform.Helpers;
 using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Models;
 
+[assembly: InternalsVisibleTo("starskytest")]
 namespace starsky.feature.rename.Services
 {
     public class RenameService
@@ -29,60 +31,9 @@ namespace starsky.feature.rename.Services
 		{
 			// -- param name="addDirectoryIfNotExist">true = create an directory if an parent directory is missing</param>
 
-			var inputFileSubPaths = PathHelper.SplitInputFilePaths(f);
-			var toFileSubPaths = PathHelper.SplitInputFilePaths(to);
-			
-			// check for the same input
-			if ( inputFileSubPaths.SequenceEqual(toFileSubPaths) )
-			{
-				return new List<FileIndexItem>{new FileIndexItem
-				{
-					Status = FileIndexItem.ExifStatus.OperationNotSupported
-				}};
-			}
-			
-			// the result list
-			var fileIndexResultsList = new List<FileIndexItem>();
-			
-			for (var i = 0; i < inputFileSubPaths.Length; i++)
-			{
-				var inputFileSubPath = PathHelper.RemoveLatestSlash(inputFileSubPaths[i]);
-				inputFileSubPaths[i] = PathHelper.PrefixDbSlash(inputFileSubPath);
+			var ((inputFileSubPaths, toFileSubPaths), fileIndexResultsList) = 
+				InputOutputSubPathsPreflight(f, to, collections);
 
-				var detailView = _query.SingleItem(inputFileSubPaths[i], null, collections, false);
-				if (detailView == null) inputFileSubPaths[i] = null;
-			}
-			
-			// To check if the file/or folder has a unique name (in database)
-			for (var i = 0; i < toFileSubPaths.Length; i++)
-			{
-				var toFileSubPath = PathHelper.RemoveLatestSlash(toFileSubPaths[i]);
-				toFileSubPaths[i] = PathHelper.PrefixDbSlash(toFileSubPath);
-
-				var detailView = _query.SingleItem(toFileSubPaths[i], null, collections, false);
-				
-				// skip for files
-				if ( detailView == null) continue;
-				// dirs are mergable (isdir=false)
-				if (detailView.FileIndexItem.IsDirectory == false) toFileSubPaths[i] = null;
-			}
-			
-			// Remove null from list
-			toFileSubPaths = toFileSubPaths.Where(p => p != null).ToArray();
-			inputFileSubPaths = inputFileSubPaths.Where(p => p != null).ToArray();
-			
-			// Check if two list are the same lenght - Change this in the future BadRequest("f != to")
-			if (toFileSubPaths.Length != inputFileSubPaths.Length || 
-				toFileSubPaths.Length == 0 || inputFileSubPaths.Length == 0) 
-			{ 
-				// files that not exist
-				fileIndexResultsList.Add(new FileIndexItem
-				{
-					Status = FileIndexItem.ExifStatus.NotFoundNotInIndex
-				});
-				return fileIndexResultsList;
-	        }
-			
 			for (var i = 0; i < toFileSubPaths.Length; i++)
 			{
 				// options
@@ -146,11 +97,7 @@ namespace starsky.feature.rename.Services
 				else if ( inputFileFolderStatus == FolderOrFileModel.FolderOrFileTypeList.File
 				          && toFileFolderStatus == FolderOrFileModel.FolderOrFileTypeList.Folder )
 				{
-					// Needed to create SetFilePath() for item that is copied, not the folder
-					// no double slash when moving to root folder
-					toFileSubPath = toFileSubPath == "/" ? $"/{FilenamesHelper.GetFileName(inputFileSubPath)}" 
-						: $"{toFileSubPath}/{FilenamesHelper.GetFileName(inputFileSubPath)}";
-					
+					toFileSubPath = GetFileName(toFileSubPath, inputFileSubPath);
 					// toFileSubPath must be the to copy directory, the filename is kept the same
 					FromFileToFolder(inputFileSubPath, toFileSubPath, fileIndexResultsList);
 				} 
@@ -169,18 +116,198 @@ namespace starsky.feature.rename.Services
 	        return fileIndexResultsList;
         }
 
+		private string GetFileName(string toFileSubPath, string inputFileSubPath)
+		{
+			// Needed to create SetFilePath() for item that is copied, not the folder
+			// no double slash when moving to root folder
+			return toFileSubPath == "/" ? $"/{FilenamesHelper.GetFileName(inputFileSubPath)}" 
+				: $"{toFileSubPath}/{FilenamesHelper.GetFileName(inputFileSubPath)}";
+		}
+		
+		/// <summary>
+		/// Checks for inputs that denied the request
+		/// </summary>
+		/// <param name="f">list of filePaths in string format (dot comma separated)</param>
+		/// <param name="to">list of filePaths in string format  (dot comma separated)</param>
+		/// <param name="collections">is Collections enabled</param>
+		/// <returns>Tuple that contains two items:
+		/// item1) Tuple of the input output string - when fails this two array's has no items
+		/// item2) the list of fileIndex Items.
+		/// This contains only values when something is wrong and the request is denied</returns>
+		internal Tuple<Tuple<string[],string[]>,List<FileIndexItem>> InputOutputSubPathsPreflight
+			(string f, string to, bool collections)
+		{
+			var inputFileSubPaths = PathHelper.SplitInputFilePaths(f).ToList();
+			var toFileSubPaths = PathHelper.SplitInputFilePaths(to).ToList();
+
+			// check for the same input
+			if ( inputFileSubPaths.SequenceEqual(toFileSubPaths) )
+			{
+				return new Tuple<Tuple<string[], string[]>, List<FileIndexItem>>(
+					new Tuple<string[], string[]>(new string[0], new string[0]),
+					new List<FileIndexItem>
+					{
+						new FileIndexItem
+						{
+							Status = FileIndexItem.ExifStatus.OperationNotSupported
+						}
+					}
+				);
+			}
+			
+			// the result list
+			var fileIndexResultsList = new List<FileIndexItem>();
+			
+			for (var i = 0; i < inputFileSubPaths.Count; i++)
+			{
+				var inputFileSubPath = PathHelper.RemoveLatestSlash(inputFileSubPaths[i]);
+				inputFileSubPaths[i] = PathHelper.PrefixDbSlash(inputFileSubPath);
+
+				var detailView = _query.SingleItem(inputFileSubPaths[i], null, collections, false);
+				if ( detailView == null )
+				{
+					inputFileSubPaths[i] = null;
+				}
+			}
+			
+			// To check if the file/or folder has a unique name (in database)
+			for (var i = 0; i < toFileSubPaths.Count; i++)
+			{
+				var toFileSubPath = PathHelper.RemoveLatestSlash(toFileSubPaths[i]);
+				toFileSubPaths[i] = PathHelper.PrefixDbSlash(toFileSubPath);
+
+				// to move
+				var detailView = _query.SingleItem(toFileSubPaths[i], null, collections, false);
+				
+				// skip for files
+				if ( detailView == null )
+				{
+					// do NOT set null because you move to location that currently doesn't exist
+					// and avoid mixing up the order of files
+					continue;
+				}
+				// dirs are mergeable, when it isn't a directory
+				if ( detailView.FileIndexItem.IsDirectory == false )
+				{
+					toFileSubPaths[i] = null;
+				}
+			}
+			
+			// // Remove null from list
+			// remove both values when ONE OF those two values are null
+			for ( var i = 0; i < toFileSubPaths.Count; i++ )
+			{
+				if ( toFileSubPaths[i] != null && inputFileSubPaths[i] != null ) continue;
+				toFileSubPaths.RemoveAt(i);
+				inputFileSubPaths.RemoveAt(i);
+				fileIndexResultsList.Add(new FileIndexItem
+				{
+					Status = FileIndexItem.ExifStatus.NotFoundNotInIndex
+				});
+			}
+			
+			// Check if two list are the same Length - Change this in the future BadRequest("f != to")
+			// when moving a file that does not exist (/non-exist.jpg to /non-exist2.jpg)
+			if (toFileSubPaths.Count != inputFileSubPaths.Count || 
+			    toFileSubPaths.Count == 0 || inputFileSubPaths.Count == 0) 
+			{ 
+				// files that not exist
+				fileIndexResultsList.Add(new FileIndexItem
+				{
+					Status = FileIndexItem.ExifStatus.NotFoundNotInIndex
+				});
+				return new Tuple<Tuple<string[], string[]>, List<FileIndexItem>>(
+					new Tuple<string[], string[]>(new string[0], new string[0]), 
+					fileIndexResultsList
+				);
+			}
+			return CollectionAddPreflight(inputFileSubPaths, toFileSubPaths, fileIndexResultsList, collections);
+		}
+
+		/// <summary>
+		/// Get the collections items when preflighting
+		/// Returns as Tuple
+		/// item1: inputFileSubPaths, toFileSubPaths
+		/// item2: list of fileIndex Results (which contains only error cases)
+		/// </summary>
+		/// <param name="inputFileSubPaths">from where to copy (file or folder)</param>
+		/// <param name="toFileSubPaths">copy to (file or folder)</param>
+		/// <param name="fileIndexResultsList">results list</param>
+		/// <param name="collections">enable file collections</param>
+		/// <returns>inputFileSubPaths list, toFileSubPaths list and fileIndexResultsList</returns>
+		private Tuple<Tuple<string[], string[]>, List<FileIndexItem>> CollectionAddPreflight(
+			IReadOnlyList<string> inputFileSubPaths, IReadOnlyList<string> toFileSubPaths,
+			List<FileIndexItem> fileIndexResultsList, bool collections)
+		{
+			if ( !collections )
+			{
+				return new Tuple<Tuple<string[], string[]>, List<FileIndexItem>>(
+					new Tuple<string[], string[]>(inputFileSubPaths.ToArray(), toFileSubPaths.ToArray()), 
+					fileIndexResultsList
+				);
+			}
+			
+			var inputCollectionFileSubPaths = new List<string>();
+			var toCollectionFileSubPaths = new List<string>();
+
+			for (var i = 0; i < inputFileSubPaths.Count; i++)
+			{
+				var collectionPaths = _query.SingleItem(inputFileSubPaths[i], 
+					null, true, false).FileIndexItem.CollectionPaths;
+				inputCollectionFileSubPaths.AddRange(collectionPaths);
+				// one file could have move than 1 collections files
+				for ( var j = 0; j < collectionPaths.Count; j++ )
+				{
+					toCollectionFileSubPaths.Add(toFileSubPaths[i]);
+				}
+			}
+
+			return new Tuple<Tuple<string[], string[]>, List<FileIndexItem>>(
+				new Tuple<string[], string[]>(inputCollectionFileSubPaths.ToArray(), toCollectionFileSubPaths.ToArray()), 
+				fileIndexResultsList
+			);
+		}
+
+		/// <summary>
+		/// Move sidecar files when those exist
+		/// </summary>
+		/// <param name="inputFileSubPath">from path</param>
+		/// <param name="toFileSubPath">to path</param>
 		private void MoveSidecarFile(string inputFileSubPath, string toFileSubPath)
 		{
-			var inputFileSubPathJsonSidecarFile =
-				JsonSidecarLocation.JsonLocation(inputFileSubPath);
-			var toSidecarFile = JsonSidecarLocation.JsonLocation(toFileSubPath);
-			if ( _iStorage.ExistFile(inputFileSubPathJsonSidecarFile) )
+			// json sidecar move
+			var jsonInputFileSubPathSidecarFile = JsonSidecarLocation
+				.JsonLocation(inputFileSubPath);
+			var jsonSidecarFile = JsonSidecarLocation.JsonLocation(toFileSubPath);
+			
+			if ( _iStorage.ExistFile(jsonInputFileSubPathSidecarFile) )
 			{
-				_iStorage.FileMove(inputFileSubPathJsonSidecarFile,toSidecarFile);
+				_iStorage.FileMove(jsonInputFileSubPathSidecarFile,jsonSidecarFile);
+			}
+
+			// xmp sidecar file move
+			if ( !ExtensionRolesHelper.IsExtensionForceXmp(inputFileSubPath) )
+			{
+				return;
+			}
+			var xmpInputFileSubPathSidecarFile =  ExtensionRolesHelper
+				.ReplaceExtensionWithXmp(inputFileSubPath);
+			var xmpSidecarFile = ExtensionRolesHelper
+				.ReplaceExtensionWithXmp(toFileSubPath);
+			if ( _iStorage.ExistFile(xmpInputFileSubPathSidecarFile) )
+			{
+				_iStorage.FileMove(xmpInputFileSubPathSidecarFile, xmpSidecarFile);
 			}
 		}
 
-		private void FromFolderToFolder(string inputFileSubPath, string toFileSubPath,
+		/// <summary>
+		/// Copy from a folder to a folder
+		/// </summary>
+		/// <param name="inputFileSubPath">from path</param>
+		/// <param name="toFileSubPath">to path</param>
+		/// <param name="fileIndexItems">list of results</param>
+		/// <exception cref="ArgumentNullException">fileIndexItems is null</exception>
+		internal void FromFolderToFolder(string inputFileSubPath, string toFileSubPath,
 			List<FileIndexItem> fileIndexItems)
 		{
 			if ( fileIndexItems == null ) throw new ArgumentNullException(nameof(fileIndexItems));